+++ src/FunctionalExt/Result/Error.cs
@@ -1,3 +1,6 @@
 namespace FunctionalExt;
 
+/// <summary>
+/// Base error type. All errors returned by the Result type should derive from this.
+/// </summary>
 public abstract record Error();
