+++ src/FunctionalExt/Result/ResultExtensions.cs
@@ -1,5 +1,3 @@
-using System.Security.Cryptography;
-
 namespace FunctionalExt;
 
 public static partial class ResultExtensions
